requirements.txt
# TradingBot V5 — Python Dependencies
# Install: pip install -r requirements.txt

# Alpaca SDK
alpaca-py>=0.21.0

# Async SQLite
aiosqlite>=0.19.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0

# Data Processing
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
